Move datafile-only fixtures into test file
